+++ client/my-sites/site-settings/settings-traffic/controller.js
@@ -7,7 +7,7 @@ import React from 'react';
  * Internal dependencies
  */
 import { renderWithReduxStore } from 'lib/react-helpers';
-import TrafficMain from 'my-sites/site-settings/traffic/main';
+import TrafficMain from 'my-sites/site-settings/settings-traffic/main';
 
 export default {
 	traffic( context ) {
