+++ client/my-sites/site-settings/settings-traffic/main.jsx
@@ -40,7 +40,7 @@ const SiteSettingsTraffic = ( {
 	translate,
 	updateFields,
 } ) => (
-	<Main className="traffic__main site-settings">
+	<Main className="settings-traffic__main site-settings">
 		<DocumentHead title={ translate( 'Site Settings' ) } />
 		<SidebarNavigation />
 		<SiteSettingsNavigation site={ site } section="traffic" />
