+++ client/wordpress-com.js
@@ -110,7 +110,7 @@ const sections = [
 	{
 		name: 'settings-traffic',
 		paths: [ '/settings/traffic', '/settings/seo', '/settings/analytics' ],
-		module: 'my-sites/site-settings/traffic',
+		module: 'my-sites/site-settings/settings-traffic',
 		secondary: true,
 		group: 'sites'
 	},
